Initial work on setting the client up to save market orders locally on disk.
This also adds a flag to disable uploading market orders to a remote server. While I have hopes that everyone uploads to a central server I don't want to force it on people. People that don't want to are either going to not run it at all, fork, block the requests, or just set the ingest URL to something broken. May as well just make life easier for people and encourage development of this version.

Have not been able to test this yet as I am on lunch and there is a queue to log into the game and the test server client isn't updating.